refactor(codegen): move storage lowering helpers
